+++ src/main/resources/db/migration/V5__Create_comment_table.sql
@@ -4,7 +4,7 @@ create table comment
 	parent_id bigint not null,
 	content varchar(1024),
 	type int not null,
-	commentator int not null,
+	commentator bigint not null,
 	gmt_create bigint not null,
 	gmt_modified bigint not null,
 	like_count bigint default 0,
